created components folder in src and moved components over there
